fix(017): green build + Taskfile/kustomize/configmap wiring (Phase 3 cleanup)
loan-origination-service now compiles cleanly and is wired into the
build pipeline + AKS deployment manifests. Resolves "build deferred"
from Phases 1-3.

Build fixes (Turk):
- Extended 5 models with 29 missing fields (ProductPricing, FraudSignals,
  IncomeVerification, CreditProfile, UserInfo) — reconciled against
  data-model.md as source of truth
- Renamed WorkflowTelemetry methods to StartWorkflowActivity /
  StartStepActivity (OTEL convention)
- Stubbed Azure.AI.Projects 2.0.0-beta.2 online orchestrator with
  NotImplementedException — SDK lacks GetAIAgentAsync(name) and
  named-param CreateAgentAsync(model:). Offline orchestrator
  remains fully functional; MVP path unaffected.
- Fixed double->decimal casts

Test fixes (Livingston):
- 45 constructor calls updated for new ILogger<T> parameter
- Switched to NullLogger<T>.Instance pattern
- PolicyEvaluationTests now mock concrete CosmosPolicyRepository
- 41 tests remain Skip'd (TDD pending impl) — none deleted

Build pipeline:
- Added loan-origination-service task to tasks/Taskfile.build.yml
  (in default task sequence)

Deploy manifests:
- Created deploy/kustomize/base/loan-origination-service.yaml
  (Deployment + Service, mirrors account-service pattern,
  workload-identity, /healthz + /readyz probes)
- Registered in kustomization.yaml (resource + image entry)
- Added LOAN_ORIGINATION_SERVICE_URL and Foundry__Mode=offline
  to configmap.yaml

Istio VirtualService deferred — deploy/kustomize/base/istio/ does
not exist in this repo.

dotnet build status: 0 errors (service + tests).

+++ src/loan-origination-service/Services/EnrichmentService.cs
@@ -114,9 +114,9 @@ public class EnrichmentService
             ApplicationNo = applicationNo,
             VerifiedMonthlyIncome = verifiedIncome,
             VerificationStatus = verificationStatus,
-            EmployerMatchPct = employerMatchPct,
+            EmployerMatchPct = (decimal)employerMatchPct,
             PayrollRecordsMonths = payrollMonths,
-            IncomeVariancePct = Math.Abs(variancePct)
+            IncomeVariancePct = (decimal)Math.Abs(variancePct)
         };
 
         _logger.LogDebug("Generated income verification for {ApplicationNo}: verified=${Income}, status={Status}",
@@ -169,8 +169,8 @@ public class EnrichmentService
         var signals = new FraudSignals
         {
             ApplicationNo = applicationNo,
-            IdentityRiskScore = identityRisk,
-            DeviceRiskScore = deviceRisk,
+            IdentityRiskScore = (decimal)identityRisk,
+            DeviceRiskScore = (decimal)deviceRisk,
             AddressMismatchFlag = addressMismatch,
             SyntheticIdFlag = syntheticId,
             WatchlistHitFlag = watchlistHit,
